Function Reference

CovidSEIR.countrymodel — Function

 countrymodel(data::CountryData, priors=defaultcountrypriors(),
              ::Type{R}=Float64) where {R <: Real} = begni

Returns Turing model for single country

CovidSEIR.covidjhudata — Method

covidjhudata()

Downloads most recent JHU CSSE data on covid cases, deaths, and recoveries.

Returns a DataFrame

CovidSEIR.plotvars — Method

plotvars(simdf::DataFrames.AbstractDataFrame,
         data::CountryData;
         dayt0=Dates.Date("2020-01-21"), # one day before JHU data begins
         colors=ColorSchemes.colorschemes[:Set1_9])

Create plots showing fit and implications of simulated trajectories.

CovidSEIR.priorreport — Function

priorreport((priors=defaultcountrypriors(), T=100, population=1e7)

Create tables and figures summarizing priors.

CovidSEIR.simtrajectories — Method

simtrajectories(cc::AbstractMCMC.AbstractChains,
                     data::CountryData, ts;
                     ic=Iterators.product(StatsBase.sample(1:size(cc,1),300, replace=false),
                                          1:size(cc,3)))

Simulate trajectories based on parameter values in chain cc.

CovidSEIR.odeSEIR — Method

odeSEIR()

Sets up ODE for SEIR model with unconfirmed cases.

Returns an ODEProblem

CovidSEIR.paramvars — Method

Transfrom ODE parameters to/from vector

CovidSEIR.paramvec — Method

Transfrom ODE parameters to/from vector

CovidSEIR.systemvars — Method

Transfrom ODE variables to/from vector

CovidSEIR.systemvec — Method

Transfrom ODE variables to/from vector

Index

CovidSEIR.countrymodel

CovidSEIR.covidjhudata

CovidSEIR.odeSEIR

CovidSEIR.paramvars

CovidSEIR.paramvec

CovidSEIR.plotvars

CovidSEIR.priorreport

CovidSEIR.simtrajectories

CovidSEIR.systemvars

CovidSEIR.systemvec